+++ packages/icons/src/descriptionMap.ts
@@ -100,7 +100,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'base',
     'walletLogo',
     'coinbaseOneLogo',
-    'diamondIncentives'
+    'diamondIncentives',
+    'spot',
+    'history'
   ],
   'speciality': [
     'hiddenEye',
@@ -148,7 +150,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'dashboard',
     'application',
     'addPeople',
-    'coinbaseOne'
+    'coinbaseOne',
+    'spot',
+    'history'
   ],
   'closed': [
     'hiddenEye',
@@ -201,7 +205,8 @@ export const descriptionMap: Record<string, IconName[]> = {
     'advancedTradeProduct',
     'candlesticks',
     'orderBook',
-    'advanced'
+    'advanced',
+    'spot'
   ],
   'math': [
     'trading',
@@ -216,7 +221,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'chartBar',
     'chartVolume',
     'pyramid',
-    'continuous'
+    'continuous',
+    'portfolio',
+    'spot'
   ],
   'mean': [
     'trading',
@@ -309,7 +316,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'paperclip',
     'list',
     'document',
-    'birthcertificate'
+    'birthcertificate',
+    'orders',
+    'history'
   ],
   '📄': [
     'taxes',
@@ -317,28 +326,33 @@ export const descriptionMap: Record<string, IconName[]> = {
     'application',
     'smartContract',
     'list',
-    'document'
+    'document',
+    'orders',
+    'history'
   ],
   '📃': [
     'taxes',
     'documentation',
     'application',
     'smartContract',
-    'document'
+    'document',
+    'history'
   ],
   '📜': [
     'taxes',
     'documentation',
     'application',
     'smartContract',
-    'document'
+    'document',
+    'history'
   ],
   '📑': [
     'taxes',
     'documentation',
     'application',
     'smartContract',
-    'document'
+    'document',
+    'history'
   ],
   'switcher': [
     'appSwitcher'
@@ -1033,7 +1047,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'learningCenter',
     'predictions',
     'inviteFriends',
-    'advisor'
+    'advisor',
+    'trade',
+    'paypalLogo'
   ],
   'profile': [
     'account',
@@ -1441,7 +1457,8 @@ export const descriptionMap: Record<string, IconName[]> = {
   'items': [
     'crystalBall',
     'list',
-    'nft'
+    'nft',
+    'orders'
   ],
   '🪄': [
     'crystalBall',
@@ -1500,7 +1517,8 @@ export const descriptionMap: Record<string, IconName[]> = {
     'comment'
   ],
   'chart pie': [
-    'chartPie'
+    'chartPie',
+    'portfolio'
   ],
   'data': [
     'chartPie',
@@ -1515,14 +1533,18 @@ export const descriptionMap: Record<string, IconName[]> = {
     'dataStack',
     'beaker',
     'allocation',
-    'config'
+    'config',
+    'portfolio',
+    'spot'
   ],
   'visualization': [
     'chartPie',
     'chartCandles',
     'chartBar',
     'chartVolume',
-    'pyramid'
+    'pyramid',
+    'portfolio',
+    'spot'
   ],
   'numbers': [
     'chartPie',
@@ -1540,7 +1562,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'chartVolume',
     'pyramid',
     'step9',
-    'calculator'
+    'calculator',
+    'portfolio',
+    'spot'
   ],
   '📊': [
     'chartPie',
@@ -1549,7 +1573,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'fib',
     'chartVolume',
     'crossTrade',
-    'pyramid'
+    'pyramid',
+    'portfolio',
+    'spot'
   ],
   '📉': [
     'chartPie',
@@ -1561,7 +1587,9 @@ export const descriptionMap: Record<string, IconName[]> = {
     'chartVolume',
     'crossTrade',
     'pyramid',
-    'chartLine'
+    'chartLine',
+    'portfolio',
+    'spot'
   ],
   '📈': [
     'chartPie',
@@ -1573,10 +1601,13 @@ export const descriptionMap: Record<string, IconName[]> = {
     'chartVolume',
     'crossTrade',
     'pyramid',
-    'chartLine'
+    'chartLine',
+    'portfolio',
+    'spot'
   ],
   '🥧': [
-    'chartPie'
+    'chartPie',
+    'portfolio'
   ],
   'chart': [
     'chartCandles',
@@ -1587,10 +1618,12 @@ export const descriptionMap: Record<string, IconName[]> = {
     'candlesticks',
     'leadChart',
     'ipoChart',
-    'lineChartPoints'
+    'lineChartPoints',
+    'spot'
   ],
   'candles': [
-    'chartCandles'
+    'chartCandles',
+    'spot'
   ],
   'passive': [
     'royalty'
@@ -1705,23 +1738,27 @@ export const descriptionMap: Record<string, IconName[]> = {
   'smart': [
     'documentation',
     'application',
-    'smartContract'
+    'smartContract',
+    'history'
   ],
   'contract': [
     'documentation',
     'application',
     'smartContract',
-    'document'
+    'document',
+    'history'
   ],
   'rules': [
     'documentation',
     'application',
-    'smartContract'
+    'smartContract',
+    'history'
   ],
   'policy': [
     'documentation',
     'application',
-    'smartContract'
+    'smartContract',
+    'history'
   ],
   'list': [
     'documentation',
@@ -1729,27 +1766,33 @@ export const descriptionMap: Record<string, IconName[]> = {
     'smartContract',
     'list',
     'staggeredList',
-    'orderBook'
+    'orderBook',
+    'orders',
+    'history'
   ],
   'agreement': [
     'documentation',
     'application',
-    'smartContract'
+    'smartContract',
+    'history'
   ],
   'commitment': [
     'documentation',
     'application',
-    'smartContract'
+    'smartContract',
+    'history'
   ],
   'arrangement': [
     'documentation',
     'application',
-    'smartContract'
+    'smartContract',
+    'history'
   ],
   'settlement': [
     'documentation',
     'application',
-    'smartContract'
+    'smartContract',
+    'history'
   ],
   'settings': [
     'settings',
@@ -3597,16 +3640,20 @@ export const descriptionMap: Record<string, IconName[]> = {
   ],
   'bullet point': [
     'list',
-    'dot'
+    'dot',
+    'orders'
   ],
   'actions': [
-    'list'
+    'list',
+    'orders'
   ],
   'tasks': [
-    'list'
+    'list',
+    'orders'
   ],
   '•': [
-    'list'
+    'list',
+    'orders'
   ],
   'computer': [
     'laptop',