+++ packages/icons/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 5.24.0 (2026-09-11)
+
+### 🚀 Features
+
+- Publish icons 2026-09-11
+
+  ##### ⭐️ Added (1)
+  - history
+
+## 5.23.0 (2026-09-11)
+
+### 🚀 Features
+
+- Publish icons 2026-09-11
+
+  ##### ⭐️ Added (5)
+  - orders
+  - portfolio
+  - trade
+  - spot
+  - paypalLogo
+
 ## 5.22.0 (8/20/2026 PST)
 
 #### 🚀 Updates